// src/app/mechanic/models/client.model.ts
import { Car } from "./car.model";

export interface Client {
  id?: number;
  nombre?: string;
  nombre_visible?: boolean;
  apellidos?: string;
  apellidos_visible?: boolean;
  dni?: string;
  dni_visible?: boolean;
  sexo?: string;
  sexo_visible?: boolean;
  cumpleannos?: string;
  cumpleannos_visible?: boolean;
  fecha_de_alta?: string;
  fecha_de_alta_visible?: boolean;
  direccion?: string;
  direccion_visible?: boolean;
  localidad?: string;
  localidad_visible?: boolean;
  codigo_postal?: string;
  codigo_postal_visible?: boolean;
  observaciones?: null;
  observaciones_visible?: boolean;
  correo?: string;
  correo_visible?: boolean;
  telef_casa?: string;
  telef_casa_visible?: boolean;
  telef_movil?: string;
  telef_movil_visible?: boolean;
  telef_otro?: string;
  telef_otro_visible?: boolean;
  vinculado?: string;
  vinculado_visible?: boolean;
  tipo?: string;
  nombre_comercial?: string;
  nombre_comercial_visible?: boolean;
  cargo?: string;
  cargo_visible?: boolean;
  web?: string;
  web_visible?: boolean;
  carro_id?: number;
  carro?: Car;
  created_at?: string;
  updated_at?: string;
}